Refactor integration tests so database access utilities are centralized and can be shared
