+++ ppr-api/requirements.txt
@@ -1,5 +1,6 @@
 
 alembic==1.3.2
+datedelta==1.3
 email-validator==1.0.5 # Uvicorn complains without this.
 psycopg2-binary==2.8.4
 python-dotenv==0.10.3
